Refactors the base sharing into its own route.
