+++ src/journal-entries/journal-entries-router.js
@@ -69,28 +69,5 @@ journalEntriesRouter
             })
             .catch(next);
     });
-    
-journalEntriesRouter
-    .get('/share/json/:entry_id', (req, res, next) => {
-        const db = req.app.get('db');
-        const requestedEntryId = req.params.entry_id;
-
-        JournalEntriesService.getEntryForPermalink(db, requestedEntryId)
-            .then(entry => {
-                if(entry.length === 0) {
-                    return res.status(400).json({error: 'Entry does not exist or is private'});
-                }
-
-                return res
-                    .status(200)
-                    .json({
-                        user: entry[0].user_name,
-                        title: entry[0].title,
-                        body: entry[0].body,
-                        created: entry[0].created
-                    });
-            })
-            .catch(next);
-    });
 
 module.exports = journalEntriesRouter;
