+++ src/journal-entries/journal-entries-service.js
@@ -34,26 +34,6 @@ JournalEntriesService = {
                 entry_id: entryId,
             });
     },
-
-    getEntryForPermalink(db, entryId) {
-        return db
-            .from('entries')
-            .select([
-                'users.user_name',
-                'entries.title',
-                'entries.body',
-                'entries.created'
-            ])
-            .innerJoin(
-                'users',
-                'entries.user_id',
-                'users.id'
-            )
-            .where({
-                entry_id: entryId,
-                privacy: 1
-            });
-    }
 };
 
 module.exports = JournalEntriesService;
